refactor(): Move core types inside registry package.
