--- typescript/onchain-actions-plugins/lib/core/.prettierrc
@@ -1,9 +1,0 @@
-{
-  "singleQuote": true,
-  "trailingComma": "es5",
-  "printWidth": 100,
-  "tabWidth": 2,
-  "useTabs": false,
-  "semi": true,
-  "bracketSpacing": true
-}
