--- typescript/onchain-actions-plugins/lib/core/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "@emberai/plugin-core",
-  "version": "1.0.0",
-  "type": "module",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "private": true,
-  "scripts": {
-    "prepare": "pnpm build",
-    "build": "tsc -b",
-    "lint": "eslint src",
-    "lint:fix": "eslint src --fix",
-    "format": "prettier --write \"src/**/*.ts\""
-  },
-  "devDependencies": {
-    "@types/node": "catalog:",
-    "typescript": "catalog:"
-  },
-  "dependencies": {
-    "@gmx-io/sdk": "^1.2.1",
-    "zod": "catalog:"
-  }
-}
