--- typescript/onchain-actions-plugins/lib/core/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src",
-    "composite": true,
-    "baseUrl": "."
-  },
-  "include": ["src/**/*"]
-}
