+++ typescript/onchain-actions-plugins/registry/src/core/actions/lending.ts
