+++ typescript/onchain-actions-plugins/registry/src/core/schemas/lending.ts
