+++ typescript/onchain-actions-plugins/registry/src/core/schemas/perpetuals.ts
