+++ typescript/onchain-actions-plugins/registry/src/core/schemas/wallet.ts
